+++ portable-atomic-util/src/arc.rs
@@ -2678,8 +2678,7 @@ mod std_impls {
     // https://doc.rust-lang.org/nightly/std/sync/struct.Arc.html#impl-AsSocket-for-Arc%3CT%3E
     // Note:
     // - T: ?Sized is currently only allowed on AsFd/AsHandle: https://github.com/rust-lang/rust/pull/114655#issuecomment-1977994288
-    // - std doesn't implement AsRawHandle/AsRawSocket for Arc as of Rust 1.77.
-    #[cfg(not(portable_atomic_no_io_safety))]
+    // - std doesn't implement AsRawHandle/AsRawSocket for Arc as of Rust 1.90.
     #[cfg(unix)]
     use std::os::unix::io as fd;
     // - std::os::unix::io::AsRawFd and std::os::windows::io::{AsRawHandle, AsRawSocket} are available in all versions
@@ -2689,8 +2688,12 @@ mod std_impls {
     // - std::os::fd requires Rust 1.66 (https://github.com/rust-lang/rust/pull/98368)
     // - std::os::hermit::io::AsFd requires Rust 1.69 (https://github.com/rust-lang/rust/commit/b5fb4f3d9b1b308d59cab24ef2f9bf23dad948aa)
     // - std::os::fd for HermitOS requires Rust 1.81 (https://github.com/rust-lang/rust/pull/126346)
+    // - std::os::fd for Trusty requires Rust 1.87 (no std support before it, https://github.com/rust-lang/rust/commit/7f6ee12526700e037ef34912b2b0c628028d382c)
     // - std::os::solid::io::AsFd is unstable (solid_ext, https://github.com/rust-lang/rust/pull/115159)
     // Note: we don't implement unstable ones.
+    #[cfg(not(portable_atomic_no_io_safety))]
+    #[cfg(target_os = "trusty")]
+    use std::os::fd;
     #[cfg(not(portable_atomic_no_io_safety))]
     #[cfg(target_os = "hermit")]
     use std::os::hermit::io as fd;
@@ -2711,9 +2714,13 @@ mod std_impls {
     /// trait MyTrait: AsRawFd {}
     /// impl MyTrait for Arc<UdpSocket> {}
     /// ```
-    // AsRawFd has been stable before io_safety, but this impl was added after io_safety: https://github.com/rust-lang/rust/pull/97437
-    #[cfg(not(portable_atomic_no_io_safety))]
-    #[cfg(any(unix, target_os = "hermit", target_os = "wasi"))]
+    #[cfg(any(
+        unix,
+        all(
+            not(portable_atomic_no_io_safety),
+            any(target_os = "hermit", target_os = "trusty", target_os = "wasi"),
+        ),
+    ))]
     impl<T: fd::AsRawFd> fd::AsRawFd for Arc<T> {
         #[inline]
         fn as_raw_fd(&self) -> fd::RawFd {
@@ -2735,7 +2742,7 @@ mod std_impls {
     /// impl MyTrait for Arc<UdpSocket> {}
     /// ```
     #[cfg(not(portable_atomic_no_io_safety))]
-    #[cfg(any(unix, target_os = "hermit", target_os = "wasi"))]
+    #[cfg(any(unix, target_os = "hermit", target_os = "trusty", target_os = "wasi"))]
     impl<T: ?Sized + fd::AsFd> fd::AsFd for Arc<T> {
         #[inline]
         fn as_fd(&self) -> fd::BorrowedFd<'_> {